[lldb/Test] Modify more tests for API replay
Skip tests or parts thereof that aren't expected to work when run from a
reproducer. Also improve the doc comments in configuration.py to prevent
mistakes in the future.

+++ lldb/test/API/functionalities/breakpoint/objc/TestObjCBreakpoints.py
@@ -120,6 +120,10 @@ class TestObjCBreakpoints(TestBase):
 
         self.check_category_breakpoints()
 
+        # Stop here for reproducers. They don't capture file system changes.
+        if configuration.is_reproducer():
+            return
+
         if have_dsym:
             shutil.rmtree(exe + ".dSYM")
         self.assertEqual(subprocess.call(
